[FIX] 유저 포인트 변경 로직에 낙관적 락 적용 및 에러 핸들링, 동시성 테스트 작성
