+++ src/main/kotlin/kr/hhplus/be/server/domain/user/UserService.kt
@@ -23,21 +23,23 @@ class UserService(
 	}
 
 	@Transactional
-	fun charge(userUUID: String, chargeAmount: Int): User {
+	fun charge(userUUID: String, chargeAmount: Int): PointHistory {
 		val user = userRepository.findByUuidForUpdate(userUUID)
 			?: throw CustomException(ErrorCode.ENTITY_NOT_FOUND, "userUUID=$userUUID")
 
 		val chargePointHistory = user.charge(chargeAmount)
-		userRepository.save(chargePointHistory)
-		return userRepository.save(user)
+
+		userRepository.save(user)
+		return userRepository.save(chargePointHistory)
 	}
 
 	@Transactional
 	fun use(userUUID: String, useAmount: Int): PointHistory {
-		val user = userRepository.findByUuid(userUUID)
+		val user = userRepository.findByUuidForUpdate(userUUID)
 			?: throw CustomException(ErrorCode.ENTITY_NOT_FOUND, "userUUID=$userUUID")
 
 		val usePointHistory = user.use(useAmount)
+
 		userRepository.save(user)
 		return userRepository.save(usePointHistory)
 	}
